codesite/requirements.txt
asgiref==3.11.1
cerebras_cloud_sdk==1.67.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.3.1
cmake==3.29.6
cohere==5.14.0
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==41.0.6
cssbeautifier==1.15.4
defusedxml==0.7.1
Django==5.2.12
django-crispy-forms==2.0
django-debug-toolbar==4.4.2
django-extensions==3.2.3
django-filter==23.2
django-taggit==4.0.0
djangorestframework==3.14.0
djlint==1.36.4
EditorConfig==0.17.1
google-genai==1.70.0
gunicorn==22.0.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
importlib-metadata==6.8.0
jsbeautifier==1.15.4
json5==0.13.0
Markdown==3.7
mistralai==2.3.1
mysqlclient==2.2.8
oauthlib==3.2.2
packaging==24.2
pathspec==1.0.4
psycopg==3.3.3
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.2
regex==2026.2.28
requests==2.31.0
requests-oauthlib==1.3.1
RestrictedPython==8.0
six==1.17.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.5.5
tomli==2.0.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.0.7
whitenoise==6.9.0
zipp==3.15.0
